mogenerator 1.0.4:
[FIX] Add -[NSEntityDescription(customBaseClass) noninheritedAttributes] and update the machine templates to use it instead of plain old -[NSEntityDescription attributesByName], which returns all attributes: even those inherited. 

Best case, this resulted in duplicate code generation in sub-entities.

Worst case, it would override human-generated methods with machine-generated methods in sub-entities. For example, consider an entity model was something like this: Human -> Employee. Human has a "firstName" attribute. Our business logic overrides the firstName accessor to prefix it with "Mr. ". This is what would happen:

 _HumanMO firstName: primitiveValueForKey: wrapper
  HumanMO firstName: return [@"Mr. " stringAppendingString:[super firstName]];
 _EmployeeMO firstName: primitiveValueForKey: wrapper (duplicate code, overrides and blows away HumanMO's behavior.)