Insert a #warning if generating a setPrimitiveType: accessor
This is a suggested alternative to the fix in commit cd9809d. That fix
results in the `primitiveType` and `setPrimitiveType:` accessors being
silently suppressed. This fix still generates them (and hence doesn't
break a build), but adds a #warning line (which Xcode will flag up)
indicating that there's a problem if the app is intended for App Store
submission.