Add the ability to generate other files
This enumerates the files in the templates path and if there are files with a motemplate extension that aren't the standard human/machine/swift h/m template files, it will run mogenerator on them. For example if there is a file called `Other.m.motemplate` mogenerator will be run on it for each entity, outputting a file called EntityNameOther.m to the machine generated directory.