Changed both scalar accessors and "Value" accessors from atomic to nonatomic
For the scalar accessors, Xcode's own code generation code makes them nonatomic, so seems to make sense to match it.

For the "Value" accessors, they just add/remove boxing to/from primitive scalar types (like int32_t) to NSNumber. The accessor they call is nonatomic, so the addition of boxing doesn't seem to me to add any atomicity to anything.